Move source files in `src/headers` folder
