+++ src/dmf/dmf.hpp
@@ -2,15 +2,15 @@
 #ifndef dmf_hpp
 #define dmf_hpp
 
-#include "./Agent.hpp"
-#include "./Graph.hpp"
-#include "./Itinerary.hpp"
-#include "./Node.hpp"
-#include "./SparseMatrix.hpp"
-#include "./Street.hpp"
-#include "./utility/HashFunctions.hpp"
-#include "./utility/TypeTraits/is_node.hpp"
-#include "./utility/TypeTraits/is_street.hpp"
-#include "./utility/TypeTraits/is_numeric.hpp"
+#include "headers/Agent.hpp"
+#include "headers/Graph.hpp"
+#include "headers/Itinerary.hpp"
+#include "headers/Node.hpp"
+#include "headers/SparseMatrix.hpp"
+#include "headers/Street.hpp"
+#include "utility/HashFunctions.hpp"
+#include "utility/TypeTraits/is_node.hpp"
+#include "utility/TypeTraits/is_street.hpp"
+#include "utility/TypeTraits/is_numeric.hpp"
 
 #endif
